weremove tailNode as a variable
if we do, then we will need to update it across many of the methods. a simpler way would be to not have to track that tailNode, and then the tail() method is a bit more verbose. i feel this is better. because if we kept the tailNode variable, we will definitely need to update it in our .append method, which i wrote in the prev commit